refactor: create helper file and move shared part
