+++ sites/partners/pages/listings/[id]/applications/pending/index.tsx
@@ -19,6 +19,7 @@ import {
 import { ListingStatusBar } from "../../../../../src/listings/ListingStatusBar"
 import Layout from "../../../../../layouts"
 import { ApplicationsSideNav } from "../../../../../src/applications/ApplicationsSideNav"
+import { tableColumns, getLinkCellFormatter } from "../../../../../src/applications/helpers"
 
 const ApplicationsList = () => {
   const router = useRouter()
@@ -31,67 +32,11 @@ const ApplicationsList = () => {
   /* Data Fetching */
   const { listingDto } = useSingleListingData(listingId)
   const listingName = listingDto?.name
-  const { data: flaggedApps } = useFlaggedApplicationsList({
+  const { data: flaggedAppsData, loading: flaggedAppsLoading } = useFlaggedApplicationsList({
     listingId,
-    page: 1,
-    limit: 1,
+    page: tableOptions.pagination.currentPage,
+    limit: tableOptions.pagination.itemsPerPage,
   })
-
-  const columns = [
-    {
-      headerName: t("applications.duplicates.duplicateGroup"),
-      field: "",
-      sortable: false,
-      filter: false,
-      pinned: "left",
-      cellRenderer: "formatLinkCell",
-    },
-    {
-      headerName: t("applications.duplicates.primaryApplicant"),
-      field: "",
-      sortable: false,
-      filter: false,
-      pinned: "left",
-    },
-    {
-      headerName: t("t.rule"),
-      field: "",
-      sortable: false,
-      filter: false,
-      pinned: "left",
-    },
-    {
-      headerName: t("applications.pendingReview"),
-      field: "",
-      sortable: false,
-      filter: false,
-      pinned: "right",
-    },
-  ]
-
-  class formatLinkCell {
-    linkWithId: HTMLSpanElement
-
-    init(params) {
-      const applicationId = params.data.id
-
-      this.linkWithId = document.createElement("button")
-      this.linkWithId.classList.add("text-blue-700")
-      this.linkWithId.innerText = params.value
-
-      this.linkWithId.addEventListener("click", function () {
-        void router.push(`/application/${applicationId}`)
-      })
-    }
-
-    getGui() {
-      return this.linkWithId
-    }
-  }
-
-  const gridComponents = {
-    formatLinkCell,
-  }
 
   return (
     <Layout>
@@ -99,12 +44,14 @@ const ApplicationsList = () => {
         <title>{t("nav.siteTitlePartners")}</title>
       </Head>
 
+      {console.log(flaggedAppsData)}
+
       <NavigationHeader
         title={listingName}
         listingId={listingId}
         tabs={{
           show: true,
-          flagsQty: flaggedApps?.meta?.totalFlagged,
+          flagsQty: flaggedAppsData?.meta?.totalFlagged,
           listingLabel: t("t.listingSingle"),
           applicationsLabel: t("nav.applications"),
           flagsLabel: t("nav.flags"),
@@ -146,15 +93,15 @@ const ApplicationsList = () => {
                 setCurrentPage: tableOptions.pagination.setCurrentPage,
               }}
               config={{
-                gridComponents,
-                columns,
+                gridComponents: { formatLinkCell: getLinkCellFormatter(router) },
+                columns: tableColumns,
                 totalItemsLabel: t("applications.totalApplications"),
               }}
               data={{
-                items: [],
-                loading: false,
-                totalItems: 0,
-                totalPages: 0,
+                items: flaggedAppsData?.items,
+                loading: flaggedAppsLoading,
+                totalItems: flaggedAppsData?.meta.totalItems,
+                totalPages: flaggedAppsData?.meta.totalPages,
               }}
               search={{
                 setSearch: tableOptions.filter.setFilterValue,
